Items

Query and return one or more API items.

See the API docs for full details on the query options available for this endpoint.

gu_items(query = NULL, show_fields = "all", show_tags = "all",
  tag = NULL, from_date = NULL, to_date = NULL,
  use_date = "published", ..., verbose = TRUE, tidy = TRUE,
  tidy_style = "snake_case")

Arguments

query

A string, containing the search query, either the URL of a single item or all items listed under a given profile, tag, etc. For example, to return all articles by a given contributor, use "profile/{contributorname}", e.g. "profile/brianlogan".

show_fields

A string or character vector of fields to include in the returned data. Defaults to "all". See details for a list of options.

show_tags

A string or character vector of tags to include in the returned data. Defaults to "all". See details for a list of options.

tag

A string or character vector of tags to filter the returned data. Defaults to NULL.

from_date

Accepts character values in 'YYYY-MM-DD' format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to NULL.

to_date

Accepts character values in 'YYYY-MM-DD' format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else that can be coerced to a date with as.Date(). Defaults to NULL.

use_date

The date type to use for the from_date and to_date parameters. One of "published", "first-publication", "newspaper-edition" or "last-modified". Defaults to "published".

...

Use to pass any other parameters to the API. See the item docs for a full list of options, including those not included here.

verbose

Prints messages to console. Defaults to TRUE.

tidy

Convert variable names to snake_case, remove some "<NA>" strings. Defaults to TRUE.

tidy_style

Style to variable names with.

Value

A tibble.

Fields options

The following are the options for the show_fields parameter:

"all" Includes all the fields (default)

"trailText"

"headline"

"showInRelatedContent" Whether this content can appear in automatically generated Related Content

"body"

"lastModified"

"hasStoryPackage" Has related content selected by editors

"score" A relevance score based on the search query used

"standfirst"

"shortUrl"

"thumbnail"

"wordcount"

"commentable"

"isPremoderated" Comments will be checked by a moderator prior to publication if true.

"allowUgc" May have associated User Generated Content. This typically means the content has an associated Guardian Witness assignment which can be accessed by querying "show-references=witness-assignment", using the query parameter.

"byline"

"publication"

"internalPageCode"

"productionOffice"

"shouldHideAdverts" Adverts will not be displayed if true

"liveBloggingNow" Content is currently live blogged if true

"commentCloseDate" The date the comments have been closed

"starRating"

The following are the options for the show_tags parameter

"blog"

"contributor"

"keyword"

"newspaper-book"

"newspaper-book-section"

"publication"

"series"

"tone"

"type"

"all": The default option.

Examples

if (FALSE) {
x <- gu_content(query = "films")
}