Sections

Returns details on the sections and subsections used to organise content.

See the API docs for full details on the query options available for the sections endpoint.

gu_section(query = NULL, ..., verbose = TRUE, tidy = TRUE,
  tidy_style = "snake_case")

Arguments

query

A string, containing the search query. Defaults to NULL, which returns all available sections subject to other parameters. Supports AND, OR and NOT operators, and exact phrase queries using double quotes. E.g. '"football" OR "politics"'. Also accepts a character vector of section names and returns those sections.

...

Use to pass any other parameters to the API. See the docs for a full list of options.

verbose

Prints messages to console. Defaults to TRUE.

tidy

Convert variable names to snake_case, remove some "<NA>" strings. Defaults to TRUE.

tidy_style

Style to variable names with.

Examples

if (FALSE) {
business <- gu_section(query = "business")

foot_pol <- gu_section(query = c("politics", "business", "football"))
}